# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * SafeWatch API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface User {
  id: string;
  firebaseUid: string;
  email: string;
  /** @nullable */
  displayName?: string | null;
  role: string;
  /** @nullable */
  parentAccountId?: string | null;
  isFoundingMember: number;
}

export interface CreateUserBody {
  displayName?: string;
}

export interface Child {
  id: string;
  userId: string;
  name: string;
  birthMonth: number;
  birthYear: number;
  entertainmentMinutes: number;
  /** @nullable */
  ageRestrictionOverride?: number | null;
  favouritesAgeBypass: number;
  eveningProtectionEnabled: number;
  eveningProtectionStartHour: number;
  eveningProtectionMaxStim: number;
  /** @nullable */
  sensitivity?: string | null;
}

export interface Video {
  id: string;
  youtubeId: string;
  title: string;
  /** @nullable */
  channelId?: string | null;
  /** @nullable */
  youtubeChannelId?: string | null;
  /** @nullable */
  youtubeChannelTitle?: string | null;
  /** @nullable */
  seriesId?: string | null;
  /** @nullable */
  seriesName?: string | null;
  durationSeconds: number;
  stimulationLevel: number;
  /** @nullable */
  ageMin?: number | null;
  /** @nullable */
  ageMax?: number | null;
  /** @nullable */
  thumbnailUrl?: string | null;
  /** @nullable */
  customThumbnailUrl?: string | null;
  isPublished: number;
  isEmbeddable: number;
}

export interface PlaylistRequest {
  childIds: string[];
  sessionMinutes: number;
  taperMode: string;
  flatlineLevel: number;
}

export type PlaylistResponseReplacementCandidates = { [key: string]: Video[] };

export interface PlaylistResponse {
  playlist: Video[];
  calmingVideos: Video[];
  replacementCandidates?: PlaylistResponseReplacementCandidates;
}

export interface CreateSessionBody {
  childIds: string[];
  totalDurationSeconds: number;
  taperMode: string;
  flatlineLevel: number;
  includeWindDown: boolean;
  finishMode: string;
}

export interface Session {
  id: string;
  userId: string;
  childIds: string[];
  startTime: string;
  /** @nullable */
  endTime?: string | null;
  totalMinutesWatched?: number;
  status: string;
  taperMode: string;
  flatlineLevel: number;
  includeWindDown: number;
  finishMode: string;
}

export interface UpdateSessionBody {
  endedAt?: string;
  totalDurationSeconds?: number;
}

export interface WatchHistoryBody {
  videoId: string;
  childIds: string[];
  sessionId: string;
  durationSeconds: number;
  stimulationLevel: number;
}

export interface WatchHistoryEntry {
  id: string;
  videoId: string;
  sessionId: string;
  durationSeconds: number;
  stimulationLevel: number;
  createdAt: string;
}

export type FeedbackBodyParticipationPcts = { [key: string]: number };

export interface BehaviorRating {
  childId: string;
  behaviorRating: string;
}

export interface FeedbackBody {
  skipped: boolean;
  behaviorRatings: BehaviorRating[];
  participationPcts: FeedbackBodyParticipationPcts;
}

export interface FeedbackResult {
  success: boolean;
}

export interface InviteInfo {
  email?: string;
  parentName?: string;
}

export interface AcceptInviteBody {
  token: string;
}