# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * SafeWatch API specification
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary Get current user profile
 */
export const GetCurrentUserResponse = zod.object({
  id: zod.string(),
  firebaseUid: zod.string(),
  email: zod.string(),
  displayName: zod.string().nullish(),
  role: zod.string(),
  parentAccountId: zod.string().nullish(),
  isFoundingMember: zod.number(),
});

/**
 * @summary Create user on first sign-up
 */
export const CreateUserBody = zod.object({
  displayName: zod.string().optional(),
});

/**
 * @summary List children for current user
 */
export const ListChildrenResponseItem = zod.object({
  id: zod.string(),
  userId: zod.string(),
  name: zod.string(),
  birthMonth: zod.number(),
  birthYear: zod.number(),
  entertainmentMinutes: zod.number(),
  ageRestrictionOverride: zod.number().nullish(),
  favouritesAgeBypass: zod.number(),
  eveningProtectionEnabled: zod.number(),
  eveningProtectionStartHour: zod.number(),
  eveningProtectionMaxStim: zod.number(),
  sensitivity: zod.string().nullish(),
});
export const ListChildrenResponse = zod.array(ListChildrenResponseItem);

/**
 * @summary Generate a playlist for a session
 */
export const GeneratePlaylistBody = zod.object({
  childIds: zod.array(zod.string()),
  sessionMinutes: zod.number(),
  taperMode: zod.string(),
  flatlineLevel: zod.number(),
});

export const GeneratePlaylistResponse = zod.object({
  playlist: zod.array(
    zod.object({
      id: zod.string(),
      youtubeId: zod.string(),
      title: zod.string(),
      channelId: zod.string().nullish(),
      youtubeChannelId: zod.string().nullish(),
      youtubeChannelTitle: zod.string().nullish(),
      seriesId: zod.string().nullish(),
      seriesName: zod.string().nullish(),
      durationSeconds: zod.number(),
      stimulationLevel: zod.number(),
      ageMin: zod.number().nullish(),
      ageMax: zod.number().nullish(),
      thumbnailUrl: zod.string().nullish(),
      customThumbnailUrl: zod.string().nullish(),
      isPublished: zod.number(),
      isEmbeddable: zod.number(),
    }),
  ),
  calmingVideos: zod.array(
    zod.object({
      id: zod.string(),
      youtubeId: zod.string(),
      title: zod.string(),
      channelId: zod.string().nullish(),
      youtubeChannelId: zod.string().nullish(),
      youtubeChannelTitle: zod.string().nullish(),
      seriesId: zod.string().nullish(),
      seriesName: zod.string().nullish(),
      durationSeconds: zod.number(),
      stimulationLevel: zod.number(),
      ageMin: zod.number().nullish(),
      ageMax: zod.number().nullish(),
      thumbnailUrl: zod.string().nullish(),
      customThumbnailUrl: zod.string().nullish(),
      isPublished: zod.number(),
      isEmbeddable: zod.number(),
    }),
  ),
  replacementCandidates: zod
    .record(
      zod.string(),
      zod.array(
        zod.object({
          id: zod.string(),
          youtubeId: zod.string(),
          title: zod.string(),
          channelId: zod.string().nullish(),
          youtubeChannelId: zod.string().nullish(),
          youtubeChannelTitle: zod.string().nullish(),
          seriesId: zod.string().nullish(),
          seriesName: zod.string().nullish(),
          durationSeconds: zod.number(),
          stimulationLevel: zod.number(),
          ageMin: zod.number().nullish(),
          ageMax: zod.number().nullish(),
          thumbnailUrl: zod.string().nullish(),
          customThumbnailUrl: zod.string().nullish(),
          isPublished: zod.number(),
          isEmbeddable: zod.number(),
        }),
      ),
    )
    .optional(),
});

/**
 * @summary Create a new viewing session
 */
export const CreateSessionBody = zod.object({
  childIds: zod.array(zod.string()),
  totalDurationSeconds: zod.number(),
  taperMode: zod.string(),
  flatlineLevel: zod.number(),
  includeWindDown: zod.boolean(),
  finishMode: zod.string(),
});

/**
 * @summary Update/end a session
 */
export const UpdateSessionParams = zod.object({
  id: zod.coerce.string(),
});

export const UpdateSessionBody = zod.object({
  endedAt: zod.date().optional(),
  totalDurationSeconds: zod.number().optional(),
});

export const UpdateSessionResponse = zod.object({
  id: zod.string(),
  userId: zod.string(),
  childIds: zod.array(zod.string()),
  startTime: zod.date(),
  endTime: zod.date().nullish(),
  totalMinutesWatched: zod.number().optional(),
  status: zod.string(),
  taperMode: zod.string(),
  flatlineLevel: zod.number(),
  includeWindDown: zod.number(),
  finishMode: zod.string(),
});

/**
 * @summary Record a video watch entry
 */
export const RecordWatchHistoryBody = zod.object({
  videoId: zod.string(),
  childIds: zod.array(zod.string()),
  sessionId: zod.string(),
  durationSeconds: zod.number(),
  stimulationLevel: zod.number(),
});

/**
 * @summary Submit session feedback
 */
export const SubmitFeedbackParams = zod.object({
  sessionId: zod.coerce.string(),
});

export const SubmitFeedbackBody = zod.object({
  skipped: zod.boolean(),
  behaviorRatings: zod.array(
    zod.object({
      childId: zod.string(),
      behaviorRating: zod.string(),
    }),
  ),
  participationPcts: zod.record(zod.string(), zod.number()),
});

export const SubmitFeedbackResponse = zod.object({
  success: zod.boolean(),
});

/**
 * @summary Validate a caregiver invite token
 */
export const ValidateInviteParams = zod.object({
  token: zod.coerce.string(),
});

export const ValidateInviteResponse = zod.object({
  email: zod.string().optional(),
  parentName: zod.string().optional(),
});

/**
 * @summary Accept a caregiver invitation
 */
export const AcceptInviteBody = zod.object({
  token: zod.string(),
});

export const AcceptInviteResponse = zod.object({
  id: zod.string(),
  firebaseUid: zod.string(),
  email: zod.string(),
  displayName: zod.string().nullish(),
  role: zod.string(),
  parentAccountId: zod.string().nullish(),
  isFoundingMember: zod.number(),
});