# lib/api-zod/src/generated/types/child.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * SafeWatch API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Child {
  id: string;
  userId: string;
  name: string;
  birthMonth: number;
  birthYear: number;
  entertainmentMinutes: number;
  /** @nullable */
  ageRestrictionOverride?: number | null;
  favouritesAgeBypass: number;
  eveningProtectionEnabled: number;
  eveningProtectionStartHour: number;
  eveningProtectionMaxStim: number;
  /** @nullable */
  sensitivity?: string | null;
}