# lib/api-zod/src/generated/types/feedbackBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * SafeWatch API specification
 * OpenAPI spec version: 0.1.0
 */
import type { BehaviorRating } from "./behaviorRating";
import type { FeedbackBodyParticipationPcts } from "./feedbackBodyParticipationPcts";

export interface FeedbackBody {
  skipped: boolean;
  behaviorRatings: BehaviorRating[];
  participationPcts: FeedbackBodyParticipationPcts;
}