# lib/api-zod/src/generated/types/playlistResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * SafeWatch API specification
 * OpenAPI spec version: 0.1.0
 */
import type { PlaylistResponseReplacementCandidates } from "./playlistResponseReplacementCandidates";
import type { Video } from "./video";

export interface PlaylistResponse {
  playlist: Video[];
  calmingVideos: Video[];
  replacementCandidates?: PlaylistResponseReplacementCandidates;
}