# lib/api-zod/src/generated/types/updateSessionBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * SafeWatch API specification
 * OpenAPI spec version: 0.1.0
 */

export interface UpdateSessionBody {
  endedAt?: Date;
  totalDurationSeconds?: number;
}