# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * SafeWatch API specification
 * OpenAPI spec version: 0.1.0
 */

export interface User {
  id: string;
  firebaseUid: string;
  email: string;
  /** @nullable */
  displayName?: string | null;
  role: string;
  /** @nullable */
  parentAccountId?: string | null;
  isFoundingMember: number;
}