# lib/api-zod/src/generated/types/video.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * SafeWatch API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Video {
  id: string;
  youtubeId: string;
  title: string;
  /** @nullable */
  channelId?: string | null;
  /** @nullable */
  youtubeChannelId?: string | null;
  /** @nullable */
  youtubeChannelTitle?: string | null;
  /** @nullable */
  seriesId?: string | null;
  /** @nullable */
  seriesName?: string | null;
  durationSeconds: number;
  stimulationLevel: number;
  /** @nullable */
  ageMin?: number | null;
  /** @nullable */
  ageMax?: number | null;
  /** @nullable */
  thumbnailUrl?: string | null;
  /** @nullable */
  customThumbnailUrl?: string | null;
  isPublished: number;
  isEmbeddable: number;
}