# lib/api-zod/src/generated/types/watchHistoryEntry.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * SafeWatch API specification
 * OpenAPI spec version: 0.1.0
 */

export interface WatchHistoryEntry {
  id: string;
  videoId: string;
  sessionId: string;
  durationSeconds: number;
  stimulationLevel: number;
  createdAt: Date;
}